+++ src/libs/actions/User.js
@@ -32,6 +32,33 @@ Onyx.connect({
 });
 
 /**
+ * Update the value of login in Onyx
+ *
+ * @param {String} newLogin
+ */
+function updateLogin(newLogin) {
+    Onyx.merge(ONYXKEYS.ACCOUNT, {login: newLogin});
+}
+
+/**
+* Update the value of password in Onyx
+*
+* @param {String} newPassword
+*/
+function updatePassword(newPassword) {
+    Onyx.merge(ONYXKEYS.ACCOUNT, {password: newPassword});
+}
+
+/**
+* Update the value of twoFactorAuthCode in Onyx
+*
+* @param {String} newTwoFactorAuthCode
+*/
+function updateTwoFactorAuthCode(newTwoFactorAuthCode) {
+    Onyx.merge(ONYXKEYS.ACCOUNT, {twoFactorAuthCode: newTwoFactorAuthCode});
+}
+
+/**
  * Changes a password for a given account
  *
  * @param {String} oldPassword
@@ -269,6 +296,9 @@ function subscribeToUserEvents() {
 }
 
 export {
+    updateLogin,
+    updatePassword,
+    updateTwoFactorAuthCode,
     changePassword,
     getBetas,
     getUserDetails,
