+++ src/pages/signin/LoginForm.js
@@ -13,13 +13,15 @@ import compose from '../../libs/compose';
 import canFocusInputOnScreenFocus from '../../libs/canFocusInputOnScreenFocus';
 import withLocalize, {withLocalizePropTypes} from '../../components/withLocalize';
 import getEmailKeyboardType from '../../libs/getEmailKeyboardType';
-import {fetchAccountDetails,updateLogin} from '../../libs/actions/Session';
+import {fetchAccountDetails} from '../../libs/actions/Session';
+import {updateLogin} from '../../libs/actions/User';
 
 const propTypes = {
     /* Onyx Props */
 
     /** The details about the account that the user is signing in with */
     account: PropTypes.shape({
+
         /** Login of the account */
         login: PropTypes.string,
